hack: slim coalesce-imports to speed up cold formatting builds
The coalesce-imports helper (run by "make fix-changed") imported
libcalico-go/lib/logutils solely to call ConfigureFormatter, a cosmetic log
prefix.  That dragged in prometheus/client_golang and its transitive tree --
254 packages, 237 of them from Prometheus alone.  On a cold build cache,
"go run ./hack/cmd/coalesce-imports" had to compile all of it (~14s) on every
invocation.

Split the custom logrus Formatter and its helpers into a new
libcalico-go/lib/logutils/logformat sub-package that depends only on the
standard library and logrus.  logutils re-exports the moved symbols as
aliases so existing callers are unchanged, and the Prometheus-using code
(BackgroundHook, Destination, etc.) stays in logutils.  coalesce-imports now
imports logformat directly, dropping its dependency tree from 254 to 90
packages.

Net effect: cold "go run" of the tool drops from ~14s to ~7s, and the warm
up-to-date check (go content-hashing the dep graph) roughly halves too.
