+++ felix/deps.txt
@@ -338,6 +338,7 @@ local:libcalico-go/lib/health
 local:libcalico-go/lib/ipam
 local:libcalico-go/lib/ipam/vmipam
 local:libcalico-go/lib/logutils
+local:libcalico-go/lib/logutils/logformat
 local:libcalico-go/lib/metricsserver
 local:libcalico-go/lib/names
 local:libcalico-go/lib/namespace
