+++ hack/deps.txt
@@ -137,6 +137,7 @@ local:libcalico-go/lib/errors
 local:libcalico-go/lib/hash
 local:libcalico-go/lib/ipam
 local:libcalico-go/lib/logutils
+local:libcalico-go/lib/logutils/logformat
 local:libcalico-go/lib/names
 local:libcalico-go/lib/namespace
 local:libcalico-go/lib/net
