+++ node/deps.txt
@@ -206,6 +206,7 @@ local:libcalico-go/lib/ipam
 local:libcalico-go/lib/ipam/vmipam
 local:libcalico-go/lib/kubevirt
 local:libcalico-go/lib/logutils
+local:libcalico-go/lib/logutils/logformat
 local:libcalico-go/lib/names
 local:libcalico-go/lib/namespace
 local:libcalico-go/lib/net
